% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adasyn.R
\name{adasyn}
\alias{adasyn}
\title{Adaptive Synthetic Minority Oversampling with ADASYN}
\usage{
adasyn(data, var, prop = 1, neighbors = 5)
}
\arguments{
\item{data}{The dataframe or tibble you wish to oversample}

\item{var}{The target class variable to be over sampled - must be a string}

\item{prop}{The proportion of minority to majority observations you wish to have after resampling - defaults to 1}
}
\description{
Adaptive Synthetic Minority Oversampling with ADASYN
}
\keyword{adasyn}
