% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_oversample.R
\name{random_oversample}
\alias{random_oversample}
\title{Randomly Oversample Minority Class in Dataframes}
\usage{
random_oversample(data, var, prop = 1)
}
\arguments{
\item{data}{The dataframe or tibble you wish to oversample}

\item{var}{The target class variable to be over sampled}

\item{prop}{The proportion of minority to majority observations you wish to have after resampling - defaults to 1}
}
\description{
Randomly Oversample Minority Class in Dataframes
}
\keyword{random_oversample}
