% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smote.R
\name{smote}
\alias{smote}
\title{Synthetic Minority Oversampling with SMOTE}
\usage{
smote(data, var, prop = 1)
}
\arguments{
\item{data}{The dataframe or tibble you wish to oversample}

\item{var}{The target class variable to be over sampled - must be a string}

\item{prop}{The proportion of minority to majority observations you wish to have after resampling - defaults to 1}
}
\description{
Synthetic Minority Oversampling with SMOTE
}
\keyword{smote}
